requirements.txt
Flask~=2.1.3
Werkzeug~=2.1.2
flask-restx~=0.5.1
Flask-CORS~=3.0.10
Flask-SQLAlchemy~=2.5.1
marshmallow~=3.17.0
PyJWT~=2.4.0
Flask-Migrate==3.1.0
pytest-cov
SQLAlchemy~=1.4.39
psycopg2-binary
gunicorn

pytest~=7.1.2
python-dotenv~=0.21.0